# ENH: Only do constrained layout at draw...