@jklymak 2021-05-14
### Benchmarks

https://github.com/matplotlib/mpl-bench

`% asv dev --bench constrained`

#### master:

```
asv dev --bench constrained
· Discovering benchmarks
· Running 2 total benchmarks (1 commits * 1 environments * 2 benchmarks)
[  0.00%] ·· Benchmarking existing-py_Users_jklymak_anaconda3_envs_matplotlibdev_bin_python
[ 25.00%] ··· constrained_layout.time_subplots                                                                                                                  ok
[ 25.00%] ··· === ==========
               n
              --- ----------
               1   55.3±0ms
               2   184±0ms
               5   833±0ms
              === ==========

[ 50.00%] ··· constrained_layout.time_subplots_colorbar                                                                                                         ok
[ 50.00%] ··· === =========
               n
              --- ---------
               1   134±0ms
               2   448±0ms
               4   1.34±0s
              === =========
```

#### This PR:

```
[ 25.00%] ··· constrained_layout.time_subplots                                                                                                                  ok
[ 25.00%] ··· === ==========
               n
              --- ----------
               1   85.5±0ms
               2   208±0ms
               5   836±0ms
              === ==========

[ 50.00%] ··· constrained_layout.time_subplots_colorbar                                                                                                         ok
[ 50.00%] ··· === =========
               n
              --- ---------
               1   120±0ms
               2   452±0ms
               4   1.35±0s
              === =========
```