@jklymak 2021-05-14
This is inspired by #19892 where we will probably move towards making layout managers more modular....