@jklymak 2021-05-15
One note - for constrained_layout, at least, the figure does need to know that it is being used before any colorbars are added so that it can use the non-gridspec colorbars.  So, in general, its not possible to strip all layout manager state to draw time.  However, this PR is still useful because it allows more layout manager state to be determined at draw time.