@jklymak 2021-06-17
### Benchmark with 30 draws per benchmark:

#### master:

```
$ asv dev --bench constrained
· Discovering benchmarks
· Running 2 total benchmarks (1 commits * 1 environments * 2 benchmarks)
[  0.00%] ·· Benchmarking existing-py_Users_jklymak_anaconda3_envs_matplotlibdev_bin_python
[ 25.00%] ··· constrained_layout.time_subplots                                ok
[ 25.00%] ··· === =========
               n           
              --- ---------
               1   868±0ms 
               2   3.27±0s 
               5   17.4±0s 
              === =========

[ 50.00%] ··· constrained_layout.time_subplots_colorbar                       ok
[ 50.00%] ··· === =========
               n           
              --- ---------
               1   2.22±0s 
               2   8.48±0s 
               4   28.9±0s 
              === =========

```

#### This PR:

```
$ asv dev --bench constrained
· Discovering benchmarks
· Running 2 total benchmarks (1 commits * 1 environments * 2 benchmarks)
[  0.00%] ·· Benchmarking existing-py_Users_jklymak_anaconda3_envs_matplotlibdev_bin_python
[ 25.00%] ··· constrained_layout.time_subplots                                ok
[ 25.00%] ··· === =========
               n           
              --- ---------
               1   928±0ms 
               2   3.31±0s 
               5   16.5±0s 
              === =========

[ 50.00%] ··· constrained_layout.time_subplots_colorbar                       ok
[ 50.00%] ··· === =========
               n           
              --- ---------
               1   2.27±0s 
               2   8.98±0s 
               4   28.1±0s 
              === =========
```

#### Same benchmark, but `constrained_layout=False`

```
asv dev --bench constrained
· Discovering benchmarks
· Running 2 total benchmarks (1 commits * 1 environments * 2 benchmarks)
[  0.00%] ·· Benchmarking existing-py_Users_jklymak_anaconda3_envs_matplotlibdev_bin_python
[ 25.00%] ··· constrained_layout.time_subplots                                ok
[ 25.00%] ··· === =========
               n           
              --- ---------
               1   404±0ms 
               2   1.57±0s 
               5   6.59±0s 
              === =========

[ 50.00%] ··· constrained_layout.time_subplots_colorbar                       ok
[ 50.00%] ··· === =========
               n           
              --- ---------
               1   925±0ms 
               2   2.95±0s 
               4   9.16±0s 
              === =========
```